+++ HealthTracker/healthtracker.client/src/components/shared/sidebar/data/sidebarLinks.ts
@@ -37,4 +37,22 @@ const Links = reactive({
     icon: "info"
   }
 })
-export default Links
+
+const AuthLinks = reactive({
+  1: {
+    name: "Register",
+    link: "/register",
+    icon: "person_add"
+  },
+  2: {
+    name: "Login",
+    link: "/login",
+    icon: "login"
+  },
+  3: {
+    name: "Logout",
+    link: "/logout",
+    icon: "logout"
+  },
+})
+export {Links, AuthLinks}
