--- HealthTracker/healthtracker.client/src/components/shared/sidebar/sidebarAuthLinks.ts
@@ -1,20 +1,0 @@
-import { reactive } from "vue";
-
-const AuthLinks = reactive({
-  1: {
-    name: "Register",
-    link: "/register",
-    icon: "person_add"
-  },
-  2: {
-    name: "Login",
-    link: "/login",
-    icon: "login"
-  },
-  3: {
-    name: "Logout",
-    link: "/logout",
-    icon: "logout"
-  },
-})
-export default AuthLinks
