+++ HealthTracker/healthtracker.client/src/components/views/diary/Diary.vue
