+++ HealthTracker/healthtracker.client/src/router/index.ts
@@ -1,13 +1,13 @@
 import {createRouter, createWebHistory} from 'vue-router'
-import Home from '../components/pages/home/Home.vue'
-import About from '../components/pages/about/About.vue'
-import Diary from '../components/pages/diary/Diary.vue'
-import TreningsPlanner from '../components/pages/treningsPlanner/TreningsPlanner.vue'
-import Health from '../components/pages/health/Health.vue'
-import Goals from '../components/pages/goals/Goals.vue'
-import Community from '../components/pages/community/Community.vue'
-import Register from '../components/pages/register/Register.vue'
-import Login from '../components/pages/login/Login.vue'
+import Home from '../components/views/home/Home.vue'
+import About from '../components/views/about/About.vue'
+import Diary from '../components/views/diary/Diary.vue'
+import TreningsPlanner from '../components/views/treningsPlanner/TreningsPlanner.vue'
+import Health from '../components/views/health/Health.vue'
+import Goals from '../components/views/goals/Goals.vue'
+import Community from '../components/views/community/Community.vue'
+import Register from '../components/views/register/Register.vue'
+import Login from '../components/views/login/Login.vue'
 
 const router = createRouter({
   history: createWebHistory(),
